+++ server/modules/proker/proker.dto.ts
@@ -1,5 +1,4 @@
 import { z } from "zod/mini";
-import { z as zo } from "zod";
 import { OPagination } from "~~/server/utils/dto/common.dto";
 import { bidangEnum, bulanEnum } from "~~/shared/enum";
 
@@ -19,16 +18,10 @@ export type TProkerCreate = z.infer<typeof OProkerCreate>;
 
 export const OProkerList = z.object({
   ...OPagination.def.shape,
-  bidang: zo.preprocess(
-    (val) => (val === "" ? undefined : val),
-    z.optional(z.enum(bidangEnum))
-  ),
-  search: z.string(),
+  bidang: z.optional(z.enum(bidangEnum)),
+  search: z.optional(z.string()),
   tahun: z.optional(z.coerce.number()),
-  bulan: zo.preprocess(
-    (val) => (val === "" ? undefined : val),
-    z.optional(z.enum(bulanEnum))
-  ),
+  bulan: z.optional(z.enum(bulanEnum)),
   mingguKe: z.optional(z.coerce.number()),
 });
